re-organize automatic AAA
