+++ HToZZ4L/cfg/run_z2l_cfg.py
@@ -7,6 +7,7 @@ import PhysicsTools.HeppyCore.framework.config as cfg
 from CMGTools.HToZZ4L.analyzers.hzz4lCore_modules_cff import * 
 from CMGTools.HToZZ4L.analyzers.hzz4lExtra_modules_cff import * 
 from CMGTools.HToZZ4L.tools.configTools import * 
+from CMGTools.RootTools.samples.autoAAAconfig import *
 
 #-------- SEQUENCE
 sequence = cfg.Sequence(hzz4lPreSequence +  [ fastSkim2L ] + hzz4lObjSequence + [
