+++ HToZZ4L/python/tools/configTools.py
@@ -1,22 +1,5 @@
 from PhysicsTools.HeppyCore.framework.heppy_loop import getHeppyOption
 
-def autoAAA(selectedComponents):
-    import re, os
-    from CMGTools.Production import changeComponentAccessMode
-    from CMGTools.Production.localityChecker import LocalityChecker
-    tier2Checker = LocalityChecker("T2_CH_CERN", datasets="/*/*/MINIAOD*")
-    for comp in selectedComponents:
-        if len(comp.files) == 0: print "ERROR, comp %s (dataset %s) has no files!" % (comp.name, getattr(comp,'dataset',None))
-        if not hasattr(comp,'dataset'): continue
-        if not re.match("/[^/]+/[^/]+/MINIAOD(SIM)?", comp.dataset): continue
-        if "/store/" not in comp.files[0]: continue
-        if re.search("/store/(group|user|cmst3)/", comp.files[0]): continue
-        if not tier2Checker.available(comp.dataset):
-            print "Dataset %s is not available, will use AAA" % comp.dataset
-            changeComponentAccessMode.convertComponent(comp, "root://cms-xrd-global.cern.ch/%s")
-            if 'X509_USER_PROXY' not in os.environ or "/afs/" not in os.environ['X509_USER_PROXY']:
-                raise RuntimeError, "X509_USER_PROXY not defined or not pointing to /afs"
-
 def redefineRunRange(selectedComponents,run_range):
     from CMGTools.HToZZ4L.samples.samples_13TeV_Fall15 import kreator
     from math import ceil
